api/requirements.txt
flask==2.0.3
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.19
werkzeug==2.0.3
PyMySQL
mysqlclient
mariadb
flask-marshmallow
marshmallow_sqlalchemy
bcrypt
pyjwt